+++ src/scalar.rs
@@ -179,6 +179,14 @@ const R3: Scalar = Scalar([
     0x6e2a_5bb9_c8db_33e9,
 ]);
 
+/// 2^-1
+const TWO_INV: Scalar = Scalar([
+    0x0000_0000_ffff_ffff,
+    0xac42_5bfd_0001_a401,
+    0xccc6_27f7_f65e_27fa,
+    0x0c12_58ac_d662_82b7,
+]);
+
 // 2^S * t = MODULUS - 1 with t odd
 const S: u32 = 32;
 
@@ -194,6 +202,23 @@ const ROOT_OF_UNITY: Scalar = Scalar([
     0x5b1b_4c80_1819_d7ec,
     0x0af5_3ae3_52a3_1e64,
     0x5bf3_adda_19e9_b27b,
+]);
+
+/// ROOT_OF_UNITY^-1
+const ROOT_OF_UNITY_INV: Scalar = Scalar([
+    0x4256_481a_dcf3_219a,
+    0x45f3_7b7f_96b6_cad3,
+    0xf9c3_f1d7_5f7a_3b27,
+    0x2d2f_c049_658a_fd43,
+]);
+
+/// GENERATOR^{2^s} where t * 2^s + 1 = q with t odd.
+/// In other words, this is a t root of unity.
+const DELTA: Scalar = Scalar([
+    0x70e3_10d3_d146_f96a,
+    0x4b64_c089_19e2_99e6,
+    0x51e1_1418_6a8b_970d,
+    0x6185_d066_27c0_67cb,
 ]);
 
 impl Default for Scalar {
@@ -341,55 +366,6 @@ impl Scalar {
         let (r7, _) = adc(0, r7, carry);
 
         Scalar::montgomery_reduce(r0, r1, r2, r3, r4, r5, r6, r7)
-    }
-
-    /// Computes the square root of this element, if it exists.
-    pub fn sqrt(&self) -> CtOption<Self> {
-        // Tonelli-Shank's algorithm for q mod 16 = 1
-        // https://eprint.iacr.org/2012/685.pdf (page 12, algorithm 5)
-
-        // w = self^((t - 1) // 2)
-        //   = self^6104339283789297388802252303364915521546564123189034618274734669823
-        let w = self.pow_vartime(&[
-            0x7fff_2dff_7fff_ffff,
-            0x04d0_ec02_a9de_d201,
-            0x94ce_bea4_199c_ec04,
-            0x0000_0000_39f6_d3a9,
-        ]);
-
-        let mut v = S;
-        let mut x = self * w;
-        let mut b = x * w;
-
-        // Initialize z as the 2^S root of unity.
-        let mut z = ROOT_OF_UNITY;
-
-        for max_v in (1..=S).rev() {
-            let mut k = 1;
-            let mut tmp = b.square();
-            let mut j_less_than_v: Choice = 1.into();
-
-            for j in 2..max_v {
-                let tmp_is_one = tmp.ct_eq(&Scalar::one());
-                let squared = Scalar::conditional_select(&tmp, &z, tmp_is_one).square();
-                tmp = Scalar::conditional_select(&squared, &tmp, tmp_is_one);
-                let new_z = Scalar::conditional_select(&z, &squared, tmp_is_one);
-                j_less_than_v &= !j.ct_eq(&v);
-                k = u32::conditional_select(&j, &k, tmp_is_one);
-                z = Scalar::conditional_select(&z, &new_z, j_less_than_v);
-            }
-
-            let result = x * z;
-            x = Scalar::conditional_select(&result, &x, b.ct_eq(&Scalar::one()));
-            z = z.square();
-            b *= z;
-            v = k;
-        }
-
-        CtOption::new(
-            x,
-            (x * x).ct_eq(self), // Only return Some if it's the square root.
-        )
     }
 
     /// Exponentiates `self` by `by`, where `by` is a
@@ -664,18 +640,13 @@ impl<'a> From<&'a Scalar> for [u8; 32] {
 }
 
 impl Field for Scalar {
+    const ZERO: Self = Self::zero();
+    const ONE: Self = Self::one();
+
     fn random(mut rng: impl RngCore) -> Self {
         let mut buf = [0; 64];
         rng.fill_bytes(&mut buf);
         Self::from_bytes_wide(&buf)
-    }
-
-    fn zero() -> Self {
-        Self::zero()
-    }
-
-    fn one() -> Self {
-        Self::one()
     }
 
     #[must_use]
@@ -692,8 +663,25 @@ impl Field for Scalar {
         self.invert()
     }
 
+    fn sqrt_ratio(num: &Self, div: &Self) -> (Choice, Self) {
+        ff::helpers::sqrt_ratio_generic(num, div)
+    }
+
     fn sqrt(&self) -> CtOption<Self> {
-        self.sqrt()
+        // (t - 1) // 2 = 6104339283789297388802252303364915521546564123189034618274734669823
+        ff::helpers::sqrt_tonelli_shanks(
+            self,
+            &[
+                0x7fff_2dff_7fff_ffff,
+                0x04d0_ec02_a9de_d201,
+                0x94ce_bea4_199c_ec04,
+                0x0000_0000_39f6_d3a9,
+            ],
+        )
+    }
+
+    fn is_zero_vartime(&self) -> bool {
+        self.0 == Self::zero().0
     }
 }
 
@@ -712,18 +700,16 @@ impl PrimeField for Scalar {
         Choice::from(self.to_bytes()[0] & 1)
     }
 
+    const MODULUS: &'static str =
+        "0x73eda753299d7d483339d80809a1d80553bda402fffe5bfeffffffff00000001";
     const NUM_BITS: u32 = MODULUS_BITS;
     const CAPACITY: u32 = Self::NUM_BITS - 1;
-
-    fn multiplicative_generator() -> Self {
-        GENERATOR
-    }
-
+    const TWO_INV: Self = TWO_INV;
+    const MULTIPLICATIVE_GENERATOR: Self = GENERATOR;
     const S: u32 = S;
-
-    fn root_of_unity() -> Self {
-        ROOT_OF_UNITY
-    }
+    const ROOT_OF_UNITY: Self = ROOT_OF_UNITY;
+    const ROOT_OF_UNITY_INV: Self = ROOT_OF_UNITY_INV;
+    const DELTA: Self = DELTA;
 }
 
 #[cfg(all(feature = "bits", not(target_pointer_width = "64")))]
@@ -783,6 +769,50 @@ where
     {
         iter.fold(Self::zero(), |acc, item| acc + item.borrow())
     }
+}
+
+impl<T> core::iter::Product<T> for Scalar
+where
+    T: core::borrow::Borrow<Scalar>,
+{
+    fn product<I>(iter: I) -> Self
+    where
+        I: Iterator<Item = T>,
+    {
+        iter.fold(Self::one(), |acc, item| acc * item.borrow())
+    }
+}
+
+#[test]
+fn test_constants() {
+    assert_eq!(
+        Scalar::MODULUS,
+        "0x73eda753299d7d483339d80809a1d80553bda402fffe5bfeffffffff00000001",
+    );
+
+    assert_eq!(Scalar::from(2) * Scalar::TWO_INV, Scalar::ONE);
+
+    assert_eq!(
+        Scalar::ROOT_OF_UNITY * Scalar::ROOT_OF_UNITY_INV,
+        Scalar::ONE,
+    );
+
+    // ROOT_OF_UNITY^{2^s} mod m == 1
+    assert_eq!(
+        Scalar::ROOT_OF_UNITY.pow(&[1u64 << Scalar::S, 0, 0, 0]),
+        Scalar::ONE,
+    );
+
+    // DELTA^{t} mod m == 1
+    assert_eq!(
+        Scalar::DELTA.pow(&[
+            0xfffe_5bfe_ffff_ffff,
+            0x09a1_d805_53bd_a402,
+            0x299d_7d48_3339_d808,
+            0x0000_0000_73ed_a753,
+        ]),
+        Scalar::ONE,
+    );
 }
 
 #[test]