+++ RELEASES.md
@@ -1,3 +1,16 @@
+# Unreleased
+
+# 0.8.0
+## Changed
+- Bumped dependencies to `ff 0.13`, `group 0.13`, `pairing 0.23`.
+
+# 0.7.1
+## Added
+- `Mul` implementations for `{G1Affine, G1Projective, G2Affine, G2Projective}`
+  with `Scalar` on the left-hand side. These enable code to be written matching
+  equations that use the additive notation `[a] P`.
+- Zeroization support for `Gt`.
+
 # 0.7.0
 
 ## Changed